requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
enum34==1.1.10
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
requests==2.24.0
rotten-tomatoes-client==0.0.3
schematics==2.1.0
urllib3==1.25.10
yarl==1.4.2
